리스트 화면, 상세화면 bg 팔레트 사용

+++ app/src/main/java/kr/pe/ssun/cokedex/data/model/UiPokemon.kt
@@ -4,6 +4,6 @@ data class UiPokemon(
     val id: Int,
     val name: String,
     val imageUrl: String,
-    val colorStart: Long = 0x00000000,
-    val colorEnd: Long = 0x00000000,
+    val colorStart: Int = 0x00000000,
+    val colorEnd: Int = 0x00000000,
 )
